move report components to reportComponents folder
