+++ app/benchmarking/components/reportComponents/enums.ts
@@ -1,0 +1,7 @@
+export enum GradingLevelsMlcEnum {
+  LOW_RISK = 'Low Risk',
+  MODERATE_LOW_RISK = 'Moderate-Low Risk',
+  MODERATE_RISK = 'Moderate Risk',
+  MODERATE_HIGH_RISK = 'Moderate-High Risk',
+  HIGH_RISK = 'High Risk',
+}
