+++ app/benchmarking/components/reportComponents/rangedBarChart.tsx
