--- app/benchmarking/components/types.ts
@@ -1,1 +1,0 @@
-export type CookbookCategoryLabels = Record<string, ('Q' | 'C' | 'T')[]>;
